refactor: Adjust model class position
